Login Functionality now works.
I am using Express JS on server. Whenever a person signs up, a hash token is created and stored in their browser. This hash token will then be used to automatically validate user login. The hash token is created using a simple algorithm, but as the key to decrypt the hash is very solid, it will be really hard to decrypt it.